Refactor index.js and use es6 modules
